refactoring woooooooooo, also fixed some spelling
