+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
         </activity>
         <activity
             android:name=".view_layer.feed.AcademicFilterActivity"
-            android:exported="true" >
+            android:exported="true">
             <intent-filter>
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
@@ -42,14 +42,14 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".create_account_use_case.CreateAccountActivity"
+            android:name=".view_layer.create_account.CreateAccountActivity"
             android:exported="true">
             <intent-filter>
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
         </activity>
         <activity
-            android:name=".login_use_case.LoginActivity"
+            android:name=".view_layer.login.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -62,7 +62,7 @@
         </activity>
         <activity
             android:name=".view_layer.feed.NoNewRecommendation"
-            android:exported="true" >
+            android:exported="true">
 
             <intent-filter>
                 <category android:name="android.intent.category.DEFAULT" />
@@ -82,10 +82,6 @@
             <intent-filter>
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
-        </activity>
-
-        <activity android:name="androidx.test.core.app.InstrumentationActivityInvoker.BootstrapActivity">
-
         </activity>
     </application>
 </manifest>
