+++ app/src/main/java/com/group80/uoftinder/interface_adapter_layer/create_account/CreateAccountViewInterface.java
@@ -1,4 +1,4 @@
-package com.group80.uoftinder.create_account_use_case;
+package com.group80.uoftinder.interface_adapter_layer.create_account;
 
 import com.group80.uoftinder.entities_layer.User;
