+++ app/src/main/java/com/group80/uoftinder/interface_adapter_layer/login/LoginPresenter.java
@@ -1,6 +1,7 @@
-package com.group80.uoftinder.login_use_case;
+package com.group80.uoftinder.interface_adapter_layer.login;
 
 import com.group80.uoftinder.entities_layer.User;
+import com.group80.uoftinder.use_case_layer.login.LoginPresenterInterface;
 
 // Interface Adapter Layer
 /**
