+++ app/src/main/java/com/group80/uoftinder/use_case_layer/feed/RecommendationInteractor.java
@@ -33,7 +33,7 @@ public class RecommendationInteractor {
         this.type = this.curUser.getUserType();
         this.usf = new UserScoreFacade(curUser);
         this.userScoreComparator = Comparator.comparing(user -> compScores.get(user));
-        // calculateCompatibilityList();
+//        calculateCompatibilityList();
         filteredCompatibilityList = new ArrayList<>();
     }
