+++ app/src/main/java/com/group80/uoftinder/view_layer/feed/RecommendationView.java
@@ -20,11 +20,11 @@ import com.group80.uoftinder.R;
 import com.group80.uoftinder.entities_layer.User;
 import com.group80.uoftinder.firebase.ProfileImagePresenter;
 import com.group80.uoftinder.firebase.ProfileImageViewInterface;
-import com.group80.uoftinder.login_use_case.LoginActivity;
+import com.group80.uoftinder.view_layer.login.LoginActivity;
 import com.group80.uoftinder.logout.LogOutInteractor;
 import com.group80.uoftinder.logout.LogOutPresenter;
 import com.group80.uoftinder.logout.LogOutViewInterface;
-import com.group80.uoftinder.presenter_layer.feed.RecommendationPresenter;
+import com.group80.uoftinder.interface_adapter_layer.feed.RecommendationPresenter;
 
 import java.io.Serializable;
 import java.util.ArrayList;
