+++ app/src/main/res/layout/activity_create_account.xml
@@ -15,16 +15,6 @@
         android:textStyle="bold" />
 
     <TextView
-        android:id="@+id/errormessage"
-        android:layout_width="200dp"
-        android:layout_height="20dp"
-        android:layout_below="@id/CreateAccountTitle"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="15dp"
-        android:text=""
-        android:textColor="@color/black" />
-
-    <TextView
         android:id="@+id/createAccountViewEmailTextView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
